python/pyproject.toml
[tool.poetry]
name = "learn-langchain"
version = "0.1.0"
description = ""
authors = ["shuntagami <shuntagami23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langserve = {extras = ["all"], version = "^0.3.1"}
openai = "^1.65.2"
anthropic = "^0.49.0"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
langchain = "^0.3.19"
langchain-community = "^0.3.18"
langchain-core = "^0.3.40"
langsmith = "^0.3.11"
python-dotenv = "^1.0.1"
langchain-openai = "^0.3.7"
langchain-anthropic = "^0.3.8"
langgraph = "^0.3.2"
tavily-python = "^0.5.1"
grandalf = "^0.8"
chainlit = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
